--- ee/vellum_ee/workflows/display/tests/workflow_serialization/generic_nodes/await_all.py
@@ -1,15 +1,0 @@
-from vellum.workflows.inputs.base import BaseInputs
-from vellum.workflows.nodes.bases.base import BaseNode
-from vellum.workflows.types.core import MergeBehavior
-
-
-class Inputs(BaseInputs):
-    input: str
-
-
-class AwaitAllGenericNode(BaseNode):
-    class Outputs(BaseNode.Outputs):
-        output = Inputs.input
-
-    class Trigger(BaseNode.Trigger):
-        merge_behavior = MergeBehavior.AWAIT_ALL
